requirements.txt
matplotlib==3.9.2
psutil==6.1.0
numpy==2.1.3
pandas==2.2.3
pyarrow==18.0.0
dask
pytorch==2.5.1
tensorboard==2.18.0